added voice commands

+++ lib/models/temp.dart
@@ -1,8 +1,8 @@
 class Temp {
   final double temperature;
-  final double humidity;
+  final int humidity;
 
-  Temp({this.temperature = 0.0, this.humidity = 0.0});
+  Temp({this.temperature = 0.0, this.humidity = 0});
 
   static Temp fromJson(Map json) =>
       Temp(temperature: json['temp'], humidity: json['hum']);
